src/boilerplate/requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
whitenoise==6.9.0
requests==2.32.4
django-jazzmin==3.0.1
drf-yasg==1.21.10
dj-database-url>=2,<3
django-environ>=0.11,<1
